refactor : file structure
