+++ examples/demo-cra/package.json
