+++ examples/demo-cra/public/favicon.ico
