+++ examples/demo-cra/public/logo512.png
