+++ examples/demo-cra/public/manifest.json
