+++ examples/demo-cra/src/components/CanvasContainer.jsx
