+++ examples/demo-cra/src/components/Color.jsx
