+++ examples/demo-cra/src/components/Pixel.jsx
