+++ examples/demo-cra/src/components/PixelType.jsx
