+++ examples/demo-cra/src/serviceWorker.js
