+++ examples/demo-cra/src/utils/averageColor.js
