+++ examples/demo-cra/src/utils/averageLastPixelColor.js
