+++ examples/demo-cra/src/utils/drawCanvasOriginal.js
