+++ examples/demo-cra/src/utils/drawHoverCanvas.js
