+++ examples/demo-cra/src/utils/drawMousemoveCanvas.js
