+++ examples/demo-cra/src/utils/pickColor.js
