+++ examples/demo-cra/src/utils/resizeImage.js
