+++ examples/demo-cra/src/utils/resizeImage.test.js
